+++ tests/tests.rs
@@ -8,7 +8,7 @@ use futures::{
     stream::{futures_unordered::FuturesUnordered, FuturesOrdered, StreamExt, TryStreamExt},
     TryFutureExt,
 };
-use pearl::{BloomProvider, Builder, Meta, Storage};
+use pearl::{BloomProvider, Builder, Meta, ReadResult, Storage};
 use rand::{seq::SliceRandom, Rng};
 use std::{
     fs,
@@ -80,7 +80,7 @@ async fn test_storage_read_write() {
     let data = b"test data string";
     write_one(&storage, 1234, data, None).await.unwrap();
     let new_data = storage.read(KeyTest::new(key)).await.unwrap();
-    assert_eq!(new_data, Bytes::copy_from_slice(data));
+    assert_eq!(new_data, ReadResult::Found(Bytes::copy_from_slice(data)));
     common::clean(storage, path).await.unwrap();
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
 }
@@ -222,7 +222,8 @@ async fn test_on_disk_index() -> Result<()> {
     assert!(path.join("test.1.blob").exists());
     info!("read {}", read_key);
     let new_data = storage.read(KeyTest::new(read_key)).await.unwrap();
-    assert_eq!(new_data, data);
+    assert!(matches!(new_data, ReadResult::Found(_)));
+    assert_eq!(new_data.unwrap(), data);
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
     common::clean(storage, path).await
 }
@@ -419,8 +420,14 @@ async fn test_read_with() {
     let data_read = storage.read(&key).await.unwrap();
     debug!("read finished");
     // data_read - last record, data_read_with - first record with "1.0" meta
+    assert!(matches!(data_read_with, ReadResult::Found(_)));
+    assert!(matches!(data_read, ReadResult::Found(_)));
+    let data_read_with = data_read_with;
     assert_ne!(data_read_with, data_read);
-    assert_eq!(data_read_with, Bytes::copy_from_slice(data1));
+    assert_eq!(
+        data_read_with,
+        ReadResult::Found(Bytes::copy_from_slice(data1))
+    );
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
 }
@@ -903,8 +910,11 @@ async fn test_mark_as_deleted_single() {
         write_one(&storage, *key, data, None).await.unwrap();
         sleep(Duration::from_millis(64)).await;
     }
-    storage.mark_all_as_deleted(&delete_key).await.unwrap();
-    assert!(!storage.contains(delete_key).await.unwrap());
+    storage.delete(&delete_key, false).await.unwrap();
+    assert!(matches!(
+        storage.contains(delete_key).await.unwrap(),
+        ReadResult::Deleted(_)
+    ));
     common::clean(storage, path).await.expect("clean failed");
     warn!("elapsed: {:.3}", now.elapsed().as_secs_f64());
 }
@@ -924,7 +934,7 @@ async fn test_mark_as_deleted_deferred_dump() {
 
     let storage = common::create_test_storage(&path, 10_000).await.unwrap();
     let update_time = std::fs::metadata(&path.join("test.0.index")).expect("metadata");
-    storage.mark_all_as_deleted(&delete_key).await.unwrap();
+    storage.delete(&delete_key, false).await.unwrap();
 
     sleep(MIN_DEFER_TIME / 2).await;
     let new_update_time = std::fs::metadata(&path.join("test.0.index")).expect("metadata");
@@ -1035,8 +1045,9 @@ async fn test_in_memory_and_disk_records_retrieval() -> Result<()> {
     );
 
     for key in keys {
-        let record = storage.read(KeyTest::new(key)).await;
-        assert_eq!(record?, records[key as usize * 2 + 1]);
+        let record = storage.read(KeyTest::new(key)).await?;
+        assert!(matches!(record, ReadResult::Found(_)));
+        assert_eq!(record.unwrap(), records[key as usize * 2 + 1]);
     }
 
     for key in keys {
@@ -1056,6 +1067,47 @@ async fn test_in_memory_and_disk_records_retrieval() -> Result<()> {
             &records[(key as usize * 2)..(key as usize * 2 + 2)]
         ));
     }
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_read_all_with_deletion_marker_delete_middle() -> Result<()> {
+    let path = common::init("delete_middle");
+    let storage = common::default_test_storage_in(path).await.unwrap();
+    let key: KeyTest = vec![0].into();
+    let data: Bytes = "test data string".repeat(16).as_bytes().to_vec().into();
+    storage.write(&key, data.clone()).await?;
+    storage.delete(&key, true).await?;
+    storage.write(&key, data.clone()).await?;
+
+    let read = storage.read_all_with_deletion_marker(&key).await?;
+
+    assert_eq!(2, read.len());
+    assert!(!read[0].is_deleted());
+    assert!(read[1].is_deleted());
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn test_read_all_with_deletion_marker_delete_middle_different_blobs() -> Result<()> {
+    let path = common::init("delete_middle_blobs");
+    let storage = common::default_test_storage_in(path).await.unwrap();
+    let key: KeyTest = vec![0].into();
+    let data: Bytes = "test data string".repeat(16).as_bytes().to_vec().into();
+    storage.write(&key, data.clone()).await?;
+    storage.try_close_active_blob().await?;
+    storage.delete(&key, false).await?;
+    storage.try_close_active_blob().await?;
+    storage.write(&key, data.clone()).await?;
+    storage.try_close_active_blob().await?;
+
+    let read = storage.read_all_with_deletion_marker(&key).await?;
+
+    assert_eq!(2, read.len());
+    assert!(!read[0].is_deleted());
+    assert!(read[1].is_deleted());
 
     Ok(())
 }